requirements.txt
# Telegram Bot
python-telegram-bot==21.0.1

# OpenAI
openai==1.12.0

# HTTP клиент с поддержкой прокси
httpx==0.27.0
httpx-socks==0.9.1

# Асинхронность
aiofiles==23.2.1

# Переменные окружения
python-dotenv==1.0.1

# База данных для хранения контекста
sqlalchemy==2.0.27
aiosqlite==0.20.0

# Логирование
loguru==0.7.2

# Валидация данных
pydantic==2.6.1
pydantic-settings==2.2.1 